// frontend/src/app/api/generated/model/frequency.ts
/**
 * RecurringTransactionsAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum Frequency {

    Monthly = 'MONTHLY',

    Quarterly = 'QUARTERLY',

    Yearly = 'YEARLY'
}