// frontend/src/app/api/generated/model/ruleType.ts
/**
 * RecurringTransactionsAPI
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum RuleType {

    Regex = 'REGEX',

    JaroWinkler = 'JARO_WINKLER',

    Amount = 'AMOUNT'
}